requirements.txt
openai>=1.13            
requests                # fetch arXiv API JSON
python-dotenv           # load .env
tiktoken                # cheap token counting
streamlit               # optional UI